requirements.txt
python-dotenv
faker
streamlit